Refactor organize-download script and handle filename conflicts.
The script was refactored into a structured `main.py` within a folder for better organization. A function to generate unique filenames was added to avoid overwriting files during the organization process. Minor improvements were made to handle edge cases like invalid directories and multipart file extensions.
